+++ storage-hbase/src/main/java/org/apache/kylin/storage/hbase/cube/v2/HBaseReadonlyStore.java
@@ -141,7 +141,7 @@ public class HBaseReadonlyStore implements IGTStore {
             }
 
             @Override
-            public int getScannedRowCount() {
+            public long getScannedRowCount() {
                 return count;
             }
         };
